관람료 페이지 분할
